+++ backend-api/src/main/java/com/ohgiraffers/backendapi/domain/community_comment/repository/CommunityCommentRepository.java
@@ -1,0 +1,11 @@
+package com.ohgiraffers.backendapi.domain.community_comment.repository;
+
+import com.ohgiraffers.backendapi.domain.community_comment.entity.CommunityComment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CommunityCommentRepository extends JpaRepository<CommunityComment, Long> {
+
+    List<CommunityComment> findByPost_PostId(Long postId);
+}
