--- backend-api/src/main/java/com/ohgiraffers/backendapi/domain/community_comments/repository/CommunityCommentRepository.java
@@ -1,11 +1,0 @@
-package com.ohgiraffers.backendapi.domain.community_comments.repository;
-
-import com.ohgiraffers.backendapi.domain.community_comments.entity.CommunityComment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface CommunityCommentRepository extends JpaRepository<CommunityComment, Long> {
-
-    List<CommunityComment> findByPost_PostId(Long postId);
-}
