--- backend-api/src/main/java/com/ohgiraffers/backendapi/domain/community_posts/dto/CommunityPostRequest.java
@@ -1,21 +1,0 @@
-package com.ohgiraffers.backendapi.domain.community_posts.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class CommunityPostRequest {
-
-    @Getter
-    @NoArgsConstructor
-    public static class Create {
-        private String title;
-        private String content;
-    }
-
-    @Getter
-    @NoArgsConstructor
-    public static class Update {
-        private String title;
-        private String content;
-    }
-}
