--- backend-api/src/main/java/com/ohgiraffers/backendapi/domain/community_posts/repository/CommunityPostRepository.java
@@ -1,7 +1,0 @@
-package com.ohgiraffers.backendapi.domain.community_posts.repository;
-
-import com.ohgiraffers.backendapi.domain.community_posts.entity.CommunityPost;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommunityPostRepository extends JpaRepository<CommunityPost, Long> {
-}
