+++ backend-api/src/main/java/com/ohgiraffers/backendapi/domain/notice/repository/NoticeRepository.java
@@ -1,0 +1,7 @@
+package com.ohgiraffers.backendapi.domain.notice.repository;
+
+import com.ohgiraffers.backendapi.domain.notice.entity.Notice;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
+}
