--- backend-api/src/main/java/com/ohgiraffers/backendapi/domain/notices/dto/NoticeRequest.java
@@ -1,18 +1,0 @@
-package com.ohgiraffers.backendapi.domain.notices.dto;
-
-import lombok.Getter;
-
-public class NoticeRequest {
-
-    @Getter
-    public static class Create {
-        private String title;
-        private String content;
-    }
-
-    @Getter
-    public static class Update {
-        private String title;
-        private String content;
-    }
-}
