--- backend-api/src/main/java/com/ohgiraffers/backendapi/domain/notices/entity/Notice.java
@@ -1,42 +1,0 @@
-package com.ohgiraffers.backendapi.domain.notices.entity;
-
-import com.ohgiraffers.backendapi.domain.user.entity.User;
-import com.ohgiraffers.backendapi.global.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "notices")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Notice extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long noticeId;
-
-    @Column(length = 100, nullable = false)
-    private String title;
-
-    @Column(columnDefinition = "TEXT", nullable = false)
-    private String content;
-
-    @Column(nullable = false)
-    @Builder.Default
-    private int views = 0;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    public void increaseViews() {
-        this.views++;
-    }
-
-    public void update(String title, String content) {
-        this.title = title;
-        this.content = content;
-    }
-}
